removed double from Fabric & added InteractiveViewer

+++ packages/fabric_repository/lib/src/constants.dart
@@ -8,6 +8,6 @@ const String purchasePrice = 'purchasePrice';
 Map<String, Map<Type, bool>> fields = {
   id: {int: false},
   title: {String: false},
-  retailPrice: {double: true},
-  purchasePrice: {double: true},
+  retailPrice: {int: true},
+  purchasePrice: {int: true},
 };
